+++ src/Indica.System.Domain/Entities/Objective.cs
@@ -5,7 +5,7 @@ namespace Indica.System.Domain.Entities
         public int IdContractProject { get; set; }
         public bool IsBasketTruck { get; set; }
         public bool IsHalfPrice { get; set; }
-        public decimal MonthlyProfitGoal { get; set; }
+        public float MonthlyProfitGoal { get; set; }
         public float FixedDivisorByMonth { get; set; }
         public int TargetOfTeamCountOnWorkday { get; set; }
         public int TargetOfTeamCountOnHoliday { get; set; }
